RavenDB-20503 Async suffix, use FileOptions.DeleteOnClose,code file locations
