+++ src/Raven.Server/Exceptions/PersistConfigurationException.cs
@@ -1,8 +1,9 @@
 ﻿using System;
+using Raven.Client.Exceptions;
 
-namespace Raven.Server.Config;
+namespace Raven.Server.Exceptions;
 
-public class PersistConfigurationException : Exception
+public class PersistConfigurationException : RavenException
 {
     public PersistConfigurationException(string message, Exception exception) 
         : base(message, exception)
